Add HQs and better district validation.
This commit adds district Headquarers to the map. These are special
cells which are assigned to a particular district and cannot be removed
from it; they sort of define the district.

During the course of adding data to the district type, some bugs were
uncovered in the district updating logic. New rules include:

- If we're drawing district 0 (the "unassigned" district) always let the
  drawing happen unless it's trying to overwrite a district HQ.
- Limit the district size as appropraite the map (e.g. 3, 5, or 7 tiles).
- Don't allow a change if it would separate a district into two
  unconnected parts.