Add a district parser to read_map.py
This commit adds a district parser which reads districts from a text
map. District HQs are denoted with a capital R or B. It also adds
assertions that the number of cells and districts on a map are odd (to
prevent ties).

For example, the district map

    Rbb
    brR

would have two districts, one with an HQ at (0,0) and one with an HQ
at (2,1). It would also throw errors because it has an even number of
cells and because it has an even number of districts.